MeetUp Events#3 and Home
Design: Was developed two logos.
MeetUpEvent#3: Changed the 'Carregando' when start the app; Arranged the default text and img in case of the error on the call of MeetUp Api; Was fixed a issues about the event's view (when there weren't future events appeared 3 past events, but when there were future event only appeared the future event), now if the length of events is less than 3, it is made a new request with 2 past events.
Home: The texts was separated on a component, was created two menu (one to home page and other to events page); The menu of home page it's in development.